% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_ActionBase.R
\docType{data}
\name{ActionBase}
\alias{ActionBase}
\title{Action Base Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ActionBase
}
\description{
Action Base Class
}
\keyword{datasets}
