% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_qp.R
\name{AddXToWeekly}
\alias{AddXToWeekly}
\title{Adds seasonal week to dataset}
\usage{
AddXToWeekly(data)
}
\arguments{
\item{data}{A data.table containing the variable \code{week}}
}
\value{
A data.table with the extra variable \code{x}
}
\description{
We often want to graph seasons. This function adds the seasonal week
to the dataset \code{data} as the variable \code{x}.
}
\examples{
library(data.table)
d <- data.table(week = 1:52)
AddXToWeekly(d)
}
