% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_simple.R
\docType{data}
\name{analyse_simple}
\alias{analyse_simple}
\title{analyse_simple}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
analyse_simple
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
