% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_covid19_areas_at_risk.R
\name{analysis_covid19_areas_at_risk}
\alias{analysis_covid19_areas_at_risk}
\title{analysis_covid19_areas_at_risk}
\usage{
analysis_covid19_areas_at_risk(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
analysis_covid19_areas_at_risk
}
