% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_covid19_nordic.R
\name{analysis_covid19_nordic}
\alias{analysis_covid19_nordic}
\title{analysis_covid19_nordic}
\usage{
analysis_covid19_nordic(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
analysis_covid19_nordic
}
