% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_normomo.R
\name{analysis_normomo}
\alias{analysis_normomo}
\title{analysis_normomo}
\usage{
analysis_normomo(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
analysis_normomo
}
