% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_simple.R
\name{analysis_simple}
\alias{analysis_simple}
\title{analysis_simple}
\usage{
analysis_simple(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
analysis_simple
}
