% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_AnalysisNormomo.R
\docType{data}
\name{AnalysisNormomo}
\alias{AnalysisNormomo}
\title{analyse_simple}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnalysisNormomo
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
