% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_AnalysisSimple.R
\docType{data}
\name{AnalysisSimple}
\alias{AnalysisSimple}
\title{analyse_simple}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnalysisSimple
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
