% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_completeness.R
\name{calculate_completeness}
\alias{calculate_completeness}
\title{Calculate completeness}
\usage{
calculate_completeness(location, x_year, x_granularity_time = "weekly",
  table = NULL)
}
\arguments{
\item{location}{location to calculate for}

\item{x_year}{to calculate for}

\item{x_granularity_time}{granularity_time}

\item{table}{db table}
}
\description{
Calculate completeness
}
