% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_completeness.R
\name{calculate_confidence_interval}
\alias{calculate_confidence_interval}
\title{Calculate Confidence Interval}
\usage{
calculate_confidence_interval(data, last_weeks = NULL)
}
\arguments{
\item{data}{the data to include}

\item{last_weeks}{if we only inlcude uncertainty for the last weeks}
}
\description{
Calculate Confidence Interval
}
