% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_env.R
\docType{data}
\name{config}
\alias{config}
\title{Flags/values to be used}
\format{
An object of class \code{environment} of length 5.
}
\usage{
config
}
\description{
Flags/values to be used
}
\keyword{datasets}
