% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mem_plots.R
\name{create_county_sheet}
\alias{create_county_sheet}
\title{create county sheet}
\usage{
create_county_sheet(conf, date)
}
\arguments{
\item{conf}{A mem model configuration object}

\item{date}{extract date}
}
\description{
create county sheet
}
