% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_input_directories.R
\name{create_input_directories}
\alias{create_input_directories}
\title{create_input_directories
This function is used to create all the necessary input directories}
\usage{
create_input_directories()
}
\description{
create_input_directories
This function is used to create all the necessary input directories
}
