% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_folders.R
\name{create_latest_folder}
\alias{create_latest_folder}
\title{Create latest folder}
\usage{
create_latest_folder(results_folder_name, date)
}
\arguments{
\item{results_folder_name}{name of the results folder}

\item{date}{the date of extraction}
}
\description{
This function copies results_folder/date til results_folder/latest
}
