% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mem_plots.R
\name{create_n_doctors_sheet}
\alias{create_n_doctors_sheet}
\title{create MEm sheet with doctors}
\usage{
create_n_doctors_sheet(conf, date)
}
\arguments{
\item{conf}{A mem model configuration object}

\item{date}{extract date}
}
\description{
create MEm sheet with doctors
}
