% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mem_plots.R
\name{create_norway_sheet}
\alias{create_norway_sheet}
\title{create norway sheet}
\usage{
create_norway_sheet(conf, date)
}
\arguments{
\item{conf}{A mem model configuration object}

\item{date}{extract date}
}
\description{
create norway sheet
}
