% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mem_plots.R
\name{create_plots}
\alias{create_plots}
\title{create MEM season plots}
\usage{
create_plots(conf, date)
}
\arguments{
\item{conf}{A mem model configuration object}

\item{date}{extract date}
}
\description{
create MEM season plots
}
