% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mem_plots.R
\name{create_region_sheet}
\alias{create_region_sheet}
\title{create MEM region sheet}
\usage{
create_region_sheet(conf, date)
}
\arguments{
\item{conf}{A mem model configuration object}

\item{date}{extract date}
}
\description{
create MEM region sheet
}
