% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covid19_daily_report.R
\name{data_covid19_daily_report}
\alias{data_covid19_daily_report}
\title{data_covid19_daily_report}
\usage{
data_covid19_daily_report(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
data_covid19_daily_report
}
