% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covid19_model.R
\name{data_covid19_model}
\alias{data_covid19_model}
\title{data_covid19_model}
\usage{
data_covid19_model(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
Get and clean model data
}
