% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covid19_nordic.R
\name{data_covid19_nordic}
\alias{data_covid19_nordic}
\title{data_covid19_nordic}
\usage{
data_covid19_nordic(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
data_covid19_nordic
}
