% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covid19_self_reporting.R
\name{data_covid19_self_reporting}
\alias{data_covid19_self_reporting}
\title{data_covid19_self_reporting}
\usage{
data_covid19_self_reporting(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
data_covid19_self_reporting
}
