% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normomo.R
\name{data_pre_normomo}
\alias{data_pre_normomo}
\title{data_pre_normomo}
\usage{
data_pre_normomo(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
data_pre_normomo
}
