% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pre_norsyss.R
\name{data_pre_norsyss}
\alias{data_pre_norsyss}
\title{data_pre_norsyss}
\usage{
data_pre_norsyss(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
data_pre_norsyss
}
