% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_DataMSIS.R
\docType{data}
\name{DataMSIS}
\alias{DataMSIS}
\title{get_MSIS_data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataMSIS
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
