% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_DataNormomo.R
\docType{data}
\name{DataNormomo}
\alias{DataNormomo}
\title{get_MSIS_data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataNormomo
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
