% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normomo.R
\name{datar_normomo_drop}
\alias{datar_normomo_drop}
\title{datar_normomo_drop}
\usage{
datar_normomo_drop(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
datar_normomo_drop
}
