% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datar_norsyss.R
\name{datar_norsyss}
\alias{datar_norsyss}
\title{datar_norsyss}
\usage{
datar_norsyss(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
datar_norsyss
}
