% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datar_norsyss_kht_email.R
\name{datar_norsyss_kht_email_drop}
\alias{datar_norsyss_kht_email_drop}
\title{datar_norsyss_kht_email_drop}
\usage{
datar_norsyss_kht_email_drop(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
datar_norsyss_kht_email_drop
}
