% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_qp.R
\name{DetermineStatus}
\alias{DetermineStatus}
\title{Normal <= threshold2
threshold2 < Medium <= threshold4
threshold4 < High}
\usage{
DetermineStatus(data)
}
\arguments{
\item{data}{A data.table containing the variables \code{n}, \code{threshold2}, and \code{threshold4}}
}
\description{
Normal <= threshold2
threshold2 < Medium <= threshold4
threshold4 < High
}
