% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rundate.R
\name{exists_rundate}
\alias{exists_rundate}
\title{greater_than_rundate
Checks to see if the rundate exists fora particular package}
\usage{
exists_rundate(pkg)
}
\arguments{
\item{pkg}{Package}
}
\description{
greater_than_rundate
Checks to see if the rundate exists fora particular package
}
