% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_qp_diagnostics.R
\name{extract_diagnostics}
\alias{extract_diagnostics}
\title{extracts diangostics data from a regression fit}
\usage{
extract_diagnostics(fit)
}
\arguments{
\item{fit}{QP regression fit}
}
\description{
extracts diangostics data from a regression fit
}
