% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_qp.R
\name{FarringtonSEinGammaSpace}
\alias{FarringtonSEinGammaSpace}
\title{Calculate Farrington SE in gamma space}
\usage{
FarringtonSEinGammaSpace(pred, phi, alpha = NULL, z = NULL,
  skewness.transform = "none")
}
\arguments{
\item{pred}{Point estimate}

\item{phi}{Dispersion}

\item{alpha}{Not used}

\item{z}{Not used}

\item{skewness.transform}{"none"/"1/2","2/3"}
}
\description{
Calculate Farrington SE in gamma space
}
