% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_qp.R
\name{FarringtonZscore}
\alias{FarringtonZscore}
\title{Farrington Z score}
\usage{
FarringtonZscore(pred, phi, alpha = NULL, z = NULL,
  skewness.transform = "none", y)
}
\arguments{
\item{pred}{Point estimate}

\item{phi}{Dispersion}

\item{alpha}{Alpha (e.g 0.05)}

\item{z}{Similar to \code{alpha} (e.g. 1.96)}

\item{skewness.transform}{"none"/"1/2","2/3"}

\item{y}{Observation}
}
\description{
Farrington Z score
}
