% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_external_api.R
\name{GenerateOutbreakListAPI}
\alias{GenerateOutbreakListAPI}
\title{Generates the external outbreak list
Generates the internal outbreak list}
\usage{
GenerateOutbreakListAPI(df, dk, useType = FALSE)
}
\arguments{
\item{df}{Data inside resYearLine.RDS}

\item{dk}{Data inside resYearLineMunicip.RDS}

\item{useType}{Use type or tag as variable name}

\item{saveFiles}{Where are you saving the outbreak lists?}
}
\description{
Generates the external outbreak list
Generates the internal outbreak list
}
