% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskManager.R
\name{get_argset_tm}
\alias{get_argset_tm}
\title{Shortcut to argset within plan within task}
\usage{
get_argset_tm(task_name, index_plan = 1, index_argset = 1)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Plan within task}

\item{index_argset}{Argset within plan}
}
\description{
Shortcut to argset within plan within task
}
