% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskManager.R
\name{get_data_tm}
\alias{get_data_tm}
\title{Shortcut to data within plan within task}
\usage{
get_data_tm(task_name, index_plan = 1, index_argset = NULL)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Plan within task}

\item{index_argset}{Not used}
}
\description{
Shortcut to data within plan within task
}
