% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_database.R
\name{get_db_connection}
\alias{get_db_connection}
\title{get_db_connection}
\usage{
get_db_connection(driver = NULL, server = NULL, port = NULL,
  user = NULL, password = NULL, db_config = config$db_config)
}
\arguments{
\item{driver}{driver}

\item{server}{server}

\item{port}{port}

\item{user}{user}

\item{password}{password}

\item{db_config}{A list containing driver, server, port, user, password}
}
\description{
get_db_connection
}
