% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_location_utilities.R
\name{get_location_name}
\alias{get_location_name}
\title{get_location_name}
\usage{
get_location_name(location_code)
}
\arguments{
\item{location_code}{a location code}
}
\description{
get_location_name
}
