% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_msis.R
\docType{data}
\name{get_MSIS_data}
\alias{get_MSIS_data}
\title{get_MSIS_data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
get_MSIS_data
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
