% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_qp_diagnostics.R
\name{get_schema_qp_diagnostics}
\alias{get_schema_qp_diagnostics}
\title{get_schema_qp_diagnostics_weekly}
\usage{
get_schema_qp_diagnostics()
}
\description{
DB schema for qp_diagnostics_weekly
}
