% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskManager.R
\name{get_schema_tm}
\alias{get_schema_tm}
\title{Shortcut to schema within task}
\usage{
get_schema_tm(task_name, index_plan = NULL, index_argset = NULL)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Not used}

\item{index_argset}{Not used}
}
\description{
Shortcut to schema within task
}
