% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskManager.R
\name{get_task_tm}
\alias{get_task_tm}
\title{Shortcut to task}
\usage{
get_task_tm(task_name, index_plan = NULL, index_argset = NULL)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Not used}

\item{index_argset}{Not used}
}
\description{
Shortcut to task
}
