% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_weather.R
\name{get_weather}
\alias{get_weather}
\title{get_weather
Gets the weather, population weighted at county and national levels}
\usage{
get_weather(impute_missing = FALSE)
}
\arguments{
\item{impute_missing}{Do you want missing data imputed?}
}
\description{
get_weather
Gets the weather, population weighted at county and national levels
}
