% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_huxtable_to_png.R
\name{huxtable_to_png}
\alias{huxtable_to_png}
\title{huxtable_to_png}
\usage{
huxtable_to_png(tab, file)
}
\arguments{
\item{tab}{Huxtable}

\item{file}{Resultant .png file}
}
\description{
huxtable_to_png
}
