% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norsyss.R
\name{IdentifyDatasets}
\alias{IdentifyDatasets}
\title{Identify the latest raw/clean datasets}
\usage{
IdentifyDatasets(raw = list.files(path("input", "norsyss"),
  "^partially_formatted_"))
}
\arguments{
\item{raw}{Folder containing raw data}

\item{clean}{Folder containing clean data}
}
\description{
Identify the latest raw/clean datasets
}
