% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_latin1_to_utf8.R
\name{latin1_to_utf8}
\alias{latin1_to_utf8}
\title{latin1_to_utf8
Takes a data.frame and converts it to utf8}
\usage{
latin1_to_utf8(df)
}
\arguments{
\item{df}{data.frame to convert}
}
\description{
latin1_to_utf8
Takes a data.frame and converts it to utf8
}
