% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_task_runner.R
\name{length_analysis_plan}
\alias{length_analysis_plan}
\title{run_with_data}
\usage{
length_analysis_plan(analysis_plan)
}
\description{
Fetches data from DB with appropriate filters
Runs analysis and writes results back to DB
}
