% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_email.R
\name{mailgun}
\alias{mailgun}
\title{mailgun}
\usage{
mailgun(subject, html = " ", to = NULL, bcc = NULL,
  attachments = NULL, inlines = NULL, include_footer = T,
  is_final = TRUE, cleanup_outlook = TRUE, ...)
}
\arguments{
\item{subject}{a}

\item{html}{a}

\item{to}{a}

\item{bcc}{a}

\item{attachments}{a}

\item{inlines}{a}

\item{include_footer}{a}

\item{is_final}{Is this a final or preliminary email?}

\item{cleanup_outlook}{Do we need to clean up problems with outlook?}

\item{...}{a}
}
\description{
mailgun
}
