% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_qp_diagnostics.R
\name{new_diagnostics_df}
\alias{new_diagnostics_df}
\title{new_diagnostics_df}
\usage{
new_diagnostics_df(n_row = 0)
}
\arguments{
\item{n_row}{determines the number of rows in the data frame.}
}
\description{
returns a data frame with the columns as needed for the diagnostics
}
