% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_analysis_normomo.R
\docType{data}
\name{r6_analysis_normomo}
\alias{r6_analysis_normomo}
\title{analyse_simple}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
r6_analysis_normomo
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
