% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_data_normomo.R
\docType{data}
\name{r6_data_normomo}
\alias{r6_data_normomo}
\title{get_MSIS_data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
r6_data_normomo
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
