% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_database.R
\docType{data}
\name{schema}
\alias{schema}
\title{schema class description}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
schema
}
\description{
schema class description
}
\keyword{datasets}
