% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pre_norsyss.R
\name{sykdomspuls_aggregate}
\alias{sykdomspuls_aggregate}
\title{sykdomspuls_aggregate}
\usage{
sykdomspuls_aggregate(date_from = "2018-01-01",
  date_to = lubridate::today(), folder = "/mount/work/projects/",
  ages = c(`0-4` = "0-4", `5-14` = "5-9", `5-14` = "10-14", `15-19` =
  "15-19", `20-29` = "20-29", `30-64` = "30-39", `30-64` = "40-49", `30-64`
  = "50-59", `30-64` = "60-64", `65-69` = "65+", `70-79` = "65+", `80+` =
  "65+"), overwrite_file = FALSE, ...)
}
\arguments{
\item{date_from}{a}

\item{date_to}{a}

\item{folder}{a}

\item{ages}{a}

\item{overwrite_file}{a}

\item{...}{a}
}
\description{
A function to extract aggregated sykdomspulsen data
}
