% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskBase.R
\docType{data}
\name{TaskBase}
\alias{TaskBase}
\title{Task Base Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaskBase
}
\description{
Task Base Class
}
\keyword{datasets}
