% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskManager.R
\docType{data}
\name{TaskManager}
\alias{TaskManager}
\title{TaskManager}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaskManager
}
\description{
An R6 Action class contains a function called 'run' that takes three arguments:
- data (list)
- arg (list)
- schema (list)
}
\details{
An action is (note: we dont explicitly create these):
- one R6 Action class
- A plnr::Plan that provide
  a) data (from the plan -- `get_data`)
  b) arguments (from the plan -- `get_argset`)
to the R6 action class

A task is:
- one R6 Action class
- A plnr::Plan that provide
  a) data (from the plan -- `get_data`)
  b) arguments (from the plan -- `get_argset`)
to the R6 action class

A TaskManager is:
- one R6 Action class
- A list of plnr::Plan's
}
\keyword{datasets}
