% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskManager.R
\name{tm_get_plans}
\alias{tm_get_plans}
\title{Shortcut to plan within task}
\usage{
tm_get_plans(task_name, index_plan = NULL, index_argset = NULL)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Not used}

\item{index_argset}{Not used}
}
\description{
Shortcut to plan within task
}
