% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_task_runner.R
\name{tm_shortcut_argset}
\alias{tm_shortcut_argset}
\title{Shortcut to argset within plan within task}
\usage{
tm_shortcut_argset(task_name, index_plan = 1, index_argset = 1)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Plan within task}

\item{index_argset}{Argset within plan}
}
\description{
Shortcut to argset within plan within task
}
