% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_shortcuts.R
\name{tm_shortcut_plan}
\alias{tm_shortcut_plan}
\title{Shortcut to plan within task}
\usage{
tm_shortcut_plan(task_name, index_plan = 1, index_argset = NULL)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Plan within task}

\item{index_argset}{Not used}
}
\description{
Shortcut to plan within task
}
