% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_task_runner.R
\name{tm_shortcut_schema}
\alias{tm_shortcut_schema}
\title{Shortcut to schema within task}
\usage{
tm_shortcut_schema(task_name, index_plan = NULL, index_argset = NULL)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Not used}

\item{index_argset}{Not used}
}
\description{
Shortcut to schema within task
}
