% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_covid19_areas_at_risk.R
\name{ui_covid19_areas_at_risk}
\alias{ui_covid19_areas_at_risk}
\title{ui_covid19_areas_at_risk}
\usage{
ui_covid19_areas_at_risk(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_covid19_areas_at_risk
}
