% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_covid19_areas_at_risk_docx.R
\name{ui_covid19_areas_at_risk_docx}
\alias{ui_covid19_areas_at_risk_docx}
\title{ui_covid19_areas_at_risk_docx}
\usage{
ui_covid19_areas_at_risk_docx(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_covid19_areas_at_risk_docx
}
