% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_covid19_areas_at_risk_utbrudd_email.R
\name{ui_covid19_areas_at_risk_utbrudd_email}
\alias{ui_covid19_areas_at_risk_utbrudd_email}
\title{ui_covid19_areas_at_risk_utbrudd_email}
\usage{
ui_covid19_areas_at_risk_utbrudd_email(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_covid19_areas_at_risk_utbrudd_email
}
