% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_covid19_nordic.R
\name{ui_covid19_nordic}
\alias{ui_covid19_nordic}
\title{ui_covid19_nordic}
\usage{
ui_covid19_nordic(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_covid19_nordic
}
