% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_threshold_plot.R
\name{ui_create_threshold_plot}
\alias{ui_create_threshold_plot}
\title{analyse_simple}
\usage{
ui_create_threshold_plot(data, output, location_code)
}
\description{
Get and clean MSIS data from msis.no
}
