% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_normomo_data_files.R
\name{ui_normomo_data_files}
\alias{ui_normomo_data_files}
\title{ui_normomo_data_files}
\usage{
ui_normomo_data_files(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_normomo_data_files
}
