% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_normomo_email_internal.R
\name{ui_normomo_email_internal}
\alias{ui_normomo_email_internal}
\title{ui_normomo_email_internal}
\usage{
ui_normomo_email_internal(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_normomo_email_internal
}
