% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_normomo_table_excess_only.R
\name{ui_normomo_table_excess_only}
\alias{ui_normomo_table_excess_only}
\title{ui_normomo_table_excess_only}
\usage{
ui_normomo_table_excess_only(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_normomo_table_excess_only
}
