% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_normomo_table_overview.R
\name{ui_normomo_table_overview}
\alias{ui_normomo_table_overview}
\title{ui_normomo_table_overview}
\usage{
ui_normomo_table_overview(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_normomo_table_overview
}
