% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_normomo_thresholds_1yr_5yr.R
\name{ui_normomo_thresholds_1yr_5yr}
\alias{ui_normomo_thresholds_1yr_5yr}
\title{ui_normomo_thresholds_1yr_5yr}
\usage{
ui_normomo_thresholds_1yr_5yr(data, argset, schema)
}
\description{
ui_normomo_thresholds_1yr_5yr
}
