% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_norsyss_kht_email.R
\name{ui_norsyss_kht_email}
\alias{ui_norsyss_kht_email}
\title{ui_norsyss_kht_email}
\usage{
ui_norsyss_kht_email(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_norsyss_kht_email
}
