% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_norsyss_pdf.R
\name{ui_norsyss_pdf}
\alias{ui_norsyss_pdf}
\title{ui_norsyss_pdf}
\usage{
ui_norsyss_pdf(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_norsyss_pdf
}
