% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_surveillance_data.R
\name{ui_surveillance_data}
\alias{ui_surveillance_data}
\title{ui_surveillance_data}
\usage{
ui_surveillance_data(data, argset, schema)
}
\arguments{
\item{data}{a}

\item{argset}{a}

\item{schema}{a}
}
\description{
ui_surveillance_data
}
