% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_threshold_plot.R
\docType{data}
\name{UICreateThresholdPlot}
\alias{UICreateThresholdPlot}
\title{analyse_simple}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UICreateThresholdPlot
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
