% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rundate.R
\name{update_rundate}
\alias{update_rundate}
\title{update_rundate
Updates the rundate db tables}
\usage{
update_rundate(task, date_run)
}
\arguments{
\item{task}{a}

\item{date_run}{the date when the analysis was run}
}
\description{
update_rundate
Updates the rundate db tables
}
