% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_external_api.R, R/ui_normomo_ssi.R
\name{ConvertConfigForAPI}
\alias{ConvertConfigForAPI}
\title{Convert the new \code{CONFIG} for use in the old external API}
\usage{
ConvertConfigForAPI(argset)

ConvertConfigForAPI(argset)
}
\description{
Convert the new \code{CONFIG} for use in the old external API

Convert the new \code{CONFIG} for use in the old external API
}
