% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_external_api.R, R/ui_normomo_ssi.R
\name{GenerateOutbreakListAPI}
\alias{GenerateOutbreakListAPI}
\title{Generates the external outbreak list
Generates the internal outbreak list}
\usage{
GenerateOutbreakListAPI(df = readRDS(fd::path("results",
  sprintf("\%s/resYearLine.RDS", LatestRawID()))),
  dk = readRDS(fd::path("results", sprintf("\%s/resYearLineMunicip.RDS",
  LatestRawID()))), saveFiles = c(fd::path("results",
  sprintf("\%s/outbreaks.RDS", LatestRawID()))), useType = FALSE)

GenerateOutbreakListAPI(df = readRDS(fd::path("results",
  sprintf("\%s/resYearLine.RDS", LatestRawID()))),
  dk = readRDS(fd::path("results", sprintf("\%s/resYearLineMunicip.RDS",
  LatestRawID()))), saveFiles = c(fd::path("results",
  sprintf("\%s/outbreaks.RDS", LatestRawID()))), useType = FALSE)
}
\arguments{
\item{df}{Data inside resYearLine.RDS}

\item{dk}{Data inside resYearLineMunicip.RDS}

\item{saveFiles}{Where are you saving the outbreak lists?}

\item{useType}{Use type or tag as variable name}

\item{df}{Data inside resYearLine.RDS}

\item{dk}{Data inside resYearLineMunicip.RDS}

\item{saveFiles}{Where are you saving the outbreak lists?}

\item{useType}{Use type or tag as variable name}
}
\description{
Generates the external outbreak list
Generates the internal outbreak list

Generates the external outbreak list
Generates the internal outbreak list
}
