% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_folders.R
\name{path}
\alias{path}
\title{This function gets the right folder for results}
\usage{
path(type = "output", results_folder_name, date, ...)
}
\arguments{
\item{results_folder_name}{name of the results folder}

\item{date}{the date of extraction}
}
\description{
This function gets the right folder for results
}
