% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_folders.R
\name{path}
\alias{path}
\title{Dashboard folders}
\usage{
path(inside = "data_raw", ..., package = config$package)
}
\arguments{
\item{inside}{where it is inside}

\item{...}{an optional path/file}

\item{package}{The name of the package}
}
\description{
This function finds folders according to the dashboard philosophy
}
