% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R, R/task_runner.R
\name{run_with_data}
\alias{run_with_data}
\title{run_analysis}
\usage{
run_with_data(table, filter, analysis_func, analysis_config,
  by_location = TRUE)

run_with_data(table, filter, analysis_func, analysis_config,
  by_location = TRUE)
}
\description{
Fetches data from DB with appropriate filters
Runs analysis and writes results back to DB

Fetches data from DB with appropriate filters
Runs analysis and writes results back to DB
}
