% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_task_runner.R
\name{run_with_data}
\alias{run_with_data}
\title{run_with_data}
\usage{
run_with_data(analysis_func, task_config, by_location = TRUE)
}
\description{
Fetches data from DB with appropriate filters
Runs analysis and writes results back to DB
}
