% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_threshold_plot.R
\docType{data}
\name{ui_create_threshold_plot}
\alias{ui_create_threshold_plot}
\title{analyse_simple}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ui_create_threshold_plot
}
\description{
Get and clean MSIS data from msis.no
}
\keyword{datasets}
